apps/momentscan-plugins/portrait-score/pyproject.toml
[project]
name = "momentscan-portrait-score"
version = "0.1.0"
description = "Portrait aesthetic scoring (CLIP-based 4-axis scoring)"
requires-python = ">=3.10"
dependencies = [
    "vpx-sdk",
    "opencv-python>=4.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
clip = ["torch>=2.0", "open-clip-torch>=2.24"]
caption = ["torch>=2.0", "open-clip-torch>=2.24", "transformers>=4.30,<5"]

[project.entry-points."visualpath.modules"]
"portrait.score" = "momentscan.portrait_score:PortraitScoreAnalyzer"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/momentscan"]

[tool.uv.sources]
vpx-sdk = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
